+++ qudtlib-model/src/main/java/io/github/qudtlib/model/FactorUnit.java
@@ -97,18 +97,6 @@ public class FactorUnit {
                 .collect(toList());
     }
 
-    public static FactorUnits normalizeFactorUnits(List<FactorUnit> factorUnits) {
-        FactorUnits ret =
-                factorUnits.stream()
-                        .map(fu -> fu.normalize())
-                        .reduce((prev, cur) -> cur.combineWith(prev))
-                        .get();
-        if (ret.isRatioOfSameUnits()) {
-            return ret;
-        }
-        return ret.reduceExponents();
-    }
-
     public static FactorUnit ofUnit(Unit unit) {
         return FactorUnit.builder().unit(unit).exponent(1).build();
     }
@@ -252,10 +240,6 @@ public class FactorUnit {
 
     private FactorUnit withExponentMultiplied(int by) {
         return FactorUnit.builder().unit(unit).exponent(this.exponent * by).build();
-    }
-
-    public FactorUnits normalize() {
-        return this.unit.normalize().pow(this.exponent);
     }
 
     public FactorUnit pow(int exponent) {